improved test

+++ test/initializeProject.test.ts
@@ -17,13 +17,32 @@ describe("checks that the app creates the necessary files to start", () => {
 		expect(fs.existsSync("./queue.initializeProject.json")).to.be.not.ok
 		expect(fs.existsSync("./failedDownloads.initializeProject.json")).to.not.be.ok
 
-		initializeProject().then(async () => {
-			expect(fs.existsSync("./queue.initializeProject.json")).to.be.ok
-			expect(fs.existsSync("./failedDownloads.initializeProject.json")).to.be.ok
+		initializeProject()
 
-			fs.unlinkSync("./queue.initializeProject.json")
-			fs.unlinkSync("./failedDownloads.initializeProject.json")
+		expect(fs.existsSync("./queue.initializeProject.json")).to.be.ok
+		expect(fs.existsSync("./failedDownloads.initializeProject.json")).to.be.ok
+
+		after(() => {
+			if (
+				fs.existsSync("./queue.initializeProject.json") ||
+				fs.existsSync("failedDownloads.initializeProject.json")
+			) {
+				fs.unlinkSync("./queue.initializeProject.json")
+				fs.unlinkSync("./failedDownloads.initializeProject.json")
+			}
 		})
+	})
+
+	it("creates the config file", () => {
+		if (fs.existsSync("./src/config.ts")) {
+			fs.unlinkSync("./src/config.ts")
+		}
+
+		expect(fs.existsSync("./src/config.ts")).to.be.not.ok
+
+		initializeProject()
+
+		expect(fs.existsSync("./src/config.ts")).to.be.ok
 
 		after(() => {
 			if (
